% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_metabolite_identification.R
\name{calc_mass_from_formula_comp_data}
\alias{calc_mass_from_formula_comp_data}
\title{calc_mass_from_formula_comp_data}
\usage{
calc_mass_from_formula_comp_data(comp_data = NULL)
}
\arguments{
\item{comp_data}{The compound database having atleast formula column}
}
\value{
The compound database with calculated mass for the compounds
}
\description{
It calculates monoisotopic mass from molecular formula within compound database
}
\examples{
calc_mass_from_formula_comp_data(comp_data)
}
