% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_spectrum_similarity.R
\name{calc_spectrum_similarity}
\alias{calc_spectrum_similarity}
\title{calc_spectrum_similarity}
\usage{
calc_spectrum_similarity(
  spec_query = NULL,
  spec_ref = NULL,
  mz_tolerence_unit = "ppm",
  mz_tolerence = 10,
  mz_range = c(0, 1600)
)
}
\arguments{
\item{spec_query}{The query spectrum dataframe with mz and intensity columns}

\item{spec_ref}{The reference spectrum dataframe with mz and intensity columns}

\item{mz_tolerence_unit}{Value of mz tolerence}

\item{mz_range}{The mz range used to calculate score}
}
\value{
The similarity score
}
\description{
It calculates the spectrum similarity score using weighted dot product algorithm.
}
\examples{
calc_spectrum_similarity <- function (spec_query, spec_ref, mz_tolerence_unit = "ppm", 
                                      mz_tolerence = 10, mz_range = c(0, 1600))
}
