% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_metabolite_identification.R
\name{calculate_mz_delta}
\alias{calculate_mz_delta}
\title{calculate_mz_delta}
\usage{
calculate_mz_delta(
  mz_source = NULL,
  mz_tolerence_unit = "ppm",
  mz_tolerence = 20
)
}
\arguments{
\item{mz_source}{mz source numeric value}

\item{mz_tolerence_unit}{The mz tolerance unit (ppm or Da)}

\item{mz_tolerence}{Value of mz tolerence}
}
\value{
A numeric value of mz delta
}
\description{
It calculates delta of particular mz value
}
\examples{
calculate_mz_delta(mz_source, mz_target, mz_tolerence_unit = "ppm", mz_tolerence = 20)
}
