% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compound_msp_to_string.R
\name{compound_msp_to_string}
\alias{compound_msp_to_string}
\title{compound_msp_to_string}
\usage{
compound_msp_to_string(compound_msp = NULL)
}
\arguments{
\item{compound_msp}{A list of elements of a single compound msp}
}
\value{
A string text of a single compound msp
}
\description{
It converts the list of elements of a single compound msp to a string text
}
\examples{
compound_msp_to_string(compound_msp)
}
