% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_adducts_rules.R
\name{create_adducts_rules}
\alias{create_adducts_rules}
\title{create_adducts_rules}
\usage{
create_adducts_rules(
  ionlist_df = NULL,
  neutralloss_df = NULL,
  neutraladdition_df = NULL,
  polarity = "positive",
  maxcharge = 3,
  mol = 3,
  nion = 2,
  nnloss = 1,
  nnadd = 1,
  nh = 2,
  lib.loc = .libPaths()
)
}
\arguments{
\item{ionlist_df}{File of known charged ions, an example is found in CAMERA/lists/ions.csv}

\item{neutralloss_df}{File of known neutral losses, an example is found in CAMERA/lists/neutralloss.csv}

\item{neutraladdition_df}{File of known adducts, an example is found in CAMERA/lists/lists/neutraladdition.csv}

\item{polarity}{The polarity of the data}

\item{maxcharge}{}

\item{mol}{Number of molecules (xM) included in the molecule charge}

\item{nion}{.}

\item{nnloss}{.}

\item{nnadd}{.}

\item{nh}{.}

\item{lib.loc}{Path to local R library}
}
\value{
A dataframe of adducts rules
}
\description{
It creates adducts rules.
}
\examples{
create_adducts_rules(ionlist_df, neutralloss_df, neutraladdition_df, polarity = 'positive')
}
