% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_xcms_object_from_elmaven.R
\name{create_xcms_object_from_elmaven}
\alias{create_xcms_object_from_elmaven}
\title{create_xcms_object_from_elmaven}
\usage{
create_xcms_object_from_elmaven(maven_output_df)
}
\arguments{
\item{maven_output_df}{The peak detailed format of El-MAVEN output}
}
\value{
xcms object
}
\description{
It create xcms object from peak detailed format of El-MAVEN
}
\examples{
create_xcms_object_from_elmaven(maven_output_df)
}
