% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_compound_msp.R
\name{get_compound_msp}
\alias{get_compound_msp}
\title{get_compound_msp}
\usage{
get_compound_msp(element_msp = NULL)
}
\arguments{
\item{element_msp}{A string of a single compound msp}
}
\value{
A list of a single compound msp
}
\description{
It converts the string text of a single compound msp into the list of msp data for a compound
}
\examples{
get_compound_msp(element_msp)
}
