% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_feature_group_representative.R
\name{get_feature_group_representative}
\alias{get_feature_group_representative}
\title{get_feature_group_representative}
\usage{
get_feature_group_representative(
  restructured_camera_df = NULL,
  polarity = "positive"
)
}
\arguments{
\item{restructured_camera_df}{The restructured camera output}

\item{polarity}{The polarity of the data}
}
\value{
The dataframe with only representative features
}
\description{
It returns a dataframe with only representative features from each feature groups.
}
\examples{
get_feature_group_representative(restructured_camera_df, polarity = "positive")
}
