% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_peak_detailed_from_emdb.R
\name{get_peak_detailed_from_emdb}
\alias{get_peak_detailed_from_emdb}
\title{get_peak_detailed_from_emdb}
\usage{
get_peak_detailed_from_emdb(emdb_path, table_name)
}
\arguments{
\item{emdb_path}{The path to emDB file}

\item{table_name}{The peak table name where desired metabolite is present}
}
\value{
ggplot object or plotly object
}
\description{
It extracts the peak detailed format of elmaven from emdb.
}
\examples{
get_peak_detailed_from_emdb(emdb_path, table_name)
}
