% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_peak_list.R
\name{get_peak_list}
\alias{get_peak_list}
\title{get_peak_list}
\usage{
get_peak_list(object, intval = "maxo")
}
\arguments{
\item{object}{xsAnnotate object}

\item{intval}{Choose intensity values. Allowed values are into, maxo, intb, intf, maxf, area, depending on the feature detection algorithm used.}
}
\value{
The dataframe of grouped peaktable
}
\description{
Extract all information from an xsAnnotate object. Returns a peaklist with annotated peaks. 
It is cloned from getPeaklist of CAMERA.
}
\examples{
getPeaklist(object, intval="maxo")
}
