% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_group_summary_from_metab_ident_format.R
\name{make_group_summary_from_metab_ident_format}
\alias{make_group_summary_from_metab_ident_format}
\title{make_group_summary_from_metab_ident_format}
\usage{
make_group_summary_from_metab_ident_format(metab_identified_df = NULL)
}
\arguments{
\item{metab_identified_df}{The output of metabolite identification step}
}
\value{
The group summary format of elmaven
}
\description{
It converts metabolites identification step or restructured camera output format to group summary format of elmaven
}
\examples{
make_group_summary_from_metab_ident_format(metab_identified_df)
}
