% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_identified_with_restructured_camera.R
\name{merge_identified_with_restructured_camera}
\alias{merge_identified_with_restructured_camera}
\title{merge_identified_with_restructured_camera}
\usage{
merge_identified_with_restructured_camera(
  metab_identified_df = NULL,
  restructured_camera_df = NULL
)
}
\arguments{
\item{metab_identified_df}{The identified metabolites dataframe}

\item{restructured_camera_df}{The restructured camera output}
}
\value{
The dataframe with merged identified and restructured camera data
}
\description{
It merges the identified metabolites data with restructured camera output data.
}
\examples{
merge_identified_with_restructured_camera(metab_identified_df, restructured_camera_df)
}
