% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_metabolite_identification.R
\name{mz_match_check}
\alias{mz_match_check}
\title{mz_match_check}
\usage{
mz_match_check(
  mz_source,
  mz_target,
  mz_tolerence_unit = "ppm",
  mz_tolerence = 20
)
}
\arguments{
\item{mz_source}{mz source value}

\item{mz_target}{mz target value}

\item{mz_tolerence_unit}{The mz tolerance unit (ppm or Da)}

\item{mz_tolerence}{Value of mz tolerence}
}
\value{
Return TRUE if mz source and mz target are within mz tolerence
}
\description{
mz_match_check
}
\examples{
mz_match_check(mz_source, mz_target, mz_tolerence_unit = "ppm", mz_tolerence = 20)
}
