% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_metabolite_identification.R
\name{mz_search_with_comp_data}
\alias{mz_search_with_comp_data}
\title{mz_search_with_comp_data}
\usage{
mz_search_with_comp_data(
  mz_source,
  comp_data,
  mz_tolerence_unit = "ppm",
  mz_tolerence = 20,
  rt_source = NULL,
  rt_tolerence = NULL
)
}
\arguments{
\item{mz_source}{mz source numeric value}

\item{comp_data}{reference compound database used for mz seraching with compulsory mzMass column}

\item{mz_tolerence_unit}{The mz tolerance unit (ppm or Da)}

\item{mz_tolerence}{Value of mz tolerence}

\item{rt_source}{rt source numeric value in minutes}

\item{rt_tolerence}{Value of rt tolerence in minutes}
}
\value{
A dataframe with mapped metabolites
}
\description{
It searches the mz within compound database
}
\examples{
mz_search_with_comp_data(mz_source, comp_data, mz_target, mz_tolerence_unit = "ppm", mz_tolerence = 20)
}
