% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_annotation_by_camera.R
\name{perform_annotation_by_camera}
\alias{perform_annotation_by_camera}
\title{perform_annotation_by_camera}
\usage{
perform_annotation_by_camera(
  xcms_object = NULL,
  polarity = NULL,
  ppm = 10,
  mzabs = 0,
  adducts_rules = NULL,
  isotope_matrix = NULL,
  cor_exp_th = 0.75,
  pval = 0.05,
  perfwhm = 0.6,
  sigma = 6,
  maxcharge = 3,
  maxiso = 4,
  minfrac = 0.5,
  multiplier = 3,
  max_peaks = 100
)
}
\arguments{
\item{xcms_object}{The xcms object having peaks information}

\item{polarity}{The polarity of the data}

\item{ppm}{General ppm error}

\item{mzabs}{General absolute error in m/z}

\item{adducts_rules}{A dataframe with different adducts rules}

\item{isotope_matrix}{four column m/z-diff and ratio Matrix, for matching isotopic peaks}

\item{cor_exp_th}{Threshold for intensity correlations across samples}

\item{pval}{p-value threshold for testing correlation of significance (0-1)}

\item{perfwhm}{percentage of FWHM (Full Width at Half Maximum) width used in groupFWHM function for grouping features}

\item{sigma}{multiplier of the standard deviation used in groupFWHM function for grouping features}

\item{maxcharge}{maximum ion charge}

\item{maxiso}{maximum number of expected isotopes (0-8)}

\item{minfrac}{The percentage number of samples, which must satisfy the C12/C13 rule for isotope annotation}

\item{multiplier}{If no ruleset is provided, calculate ruleset with max. number n of [nM+x] cluster ions}

\item{max_peaks}{How many peaks will be calculated in every thread using the parallel mode}
}
\value{
A dataframe of annotated peaks
}
\description{
It performs annotation by CAMERA using different functions of it.
}
\examples{
perform_annotation_by_camera(xcms_object, polarity = "positive", ppm = 5, mzabs = 0.01)
}
