% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_eic_from_emdb.R
\name{plot_eic_from_emdb}
\alias{plot_eic_from_emdb}
\title{plot_eic_from_emdb}
\usage{
plot_eic_from_emdb(
  emdb_path = NULL,
  table_group_id = NULL,
  table_name = NULL,
  x_label = "RT",
  y_label = "Intensity",
  title_label = "",
  legend_label = "Sample",
  x_label_size = 15,
  y_label_size = 15,
  title_label_size = 15,
  legend_label_size = 15,
  interactive = T
)
}
\arguments{
\item{emdb_path}{The path to emDB file}

\item{table_group_id}{The group id of desired metabolite}

\item{table_name}{The peak table name where desired metabolite is present}

\item{x_label}{Label x-axis}

\item{y_label}{Label x-axis}

\item{title_label}{Title of the plot}

\item{legend_label}{Label for the lagends}

\item{x_label_size}{Text size of x label}

\item{y_label_size}{Text size of y label}

\item{title_label_size}{Text size of text label}

\item{legend_label_size}{Text size of legend label}

\item{interactive}{Make plot interactive using plotly}
}
\value{
ggplot object or plotly object
}
\description{
It plots the extracted ion chromatogram (EIC) for specific groupid from emdb.
}
\examples{
plot_eic_from_emdb <- function(emdb_path, table_group_id, table_name)
}
