% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hist_elements_frequency.R
\name{plot_hist_elements_frequency}
\alias{plot_hist_elements_frequency}
\title{plot_hist_elements_frequency}
\usage{
plot_hist_elements_frequency(
  elements_vec = NULL,
  frequency_type = "by_occurrence",
  plot_title = NULL,
  xaxis_title = NULL,
  yaxis_title = NULL
)
}
\arguments{
\item{elements_vec}{The dataframe 1}

\item{frequency_type}{Show bar plot frequency by occurrence (by_occurrence) or by elements (by_elements)}

\item{plot_title}{The title of the plot}

\item{xaxis_title}{The title of the x-axis}

\item{yaxis_title}{The title of the y-axis}
}
\value{
The plotly object of bar plot
}
\description{
It plots a histogram showing the ferquency of each element
}
\examples{
plot_hist_elements_frequency(elements_vec, frequency_type = "by_occurrence")
}
