% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mass_spectrum.R
\name{plot_mass_spectrum}
\alias{plot_mass_spectrum}
\title{plot_mass_spectrum}
\usage{
plot_mass_spectrum(
  ms_spectrum = NULL,
  annotate_mz = FALSE,
  annotate_mz_cutoff = NULL,
  intensity_cutoff = NULL,
  mz_precision = NULL,
  show_rel_intensity = FALSE,
  mz_range = NULL,
  start_from_zero = TRUE,
  x_label = NULL,
  y_label = NULL,
  title_label = NULL,
  interactive = FALSE
)
}
\arguments{
\item{ms_spectrum}{The query spectrum dataframe with mz and intensity columns}

\item{annotate_mz}{Annotate mz values on spectrum, select from TRUE/FALSE}

\item{annotate_mz_cutoff}{The intensity cutoff in percentage w.r.t base peak (highest intensity peak) used to annotate mz peaks}

\item{intensity_cutoff}{The intensity cutoff in percentage w.r.t base peak (highest intensity peak) used to filter data}

\item{mz_precision}{The numeric precision of mz which is equal to the number of digits to show on mz annotation}

\item{show_rel_intensity}{Show relative intensity in percentage (TRUE) on y axis which is calculated w.r.t base peak}

\item{mz_range}{The mz range used to calculate score}

\item{start_from_zero}{Force x and y axes to start from zero (TRUE/FALSE)}

\item{x_label}{Label x-axis}

\item{y_label}{Label x-axis}

\item{title_label}{Title of the plot}

\item{interactive}{Make plot interactive using plotly}
}
\value{
ggplot object or plotly object
}
\description{
It plots the mass spectrum
}
\examples{
plot_mass_spectrum(ms_spectrum)
}
