% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spectrum_similarity.R
\name{plot_spectrum_similarity}
\alias{plot_spectrum_similarity}
\title{plot_spectrum_similarity}
\usage{
plot_spectrum_similarity(
  spec_query = NULL,
  spec_ref = NULL,
  mz_tolerence_unit = "ppm",
  mz_tolerence = 20,
  mz_range = c(0, 1600),
  interactive = FALSE,
  x_label = "mz",
  y_label = "intensity",
  title_label = ""
)
}
\arguments{
\item{spec_query}{The query spectrum dataframe with mz and intensity columns}

\item{spec_ref}{The reference spectrum dataframe with mz and intensity columns}

\item{mz_tolerence_unit}{Value of mz tolerence}

\item{mz_range}{The mz range used to calculate score}

\item{interactive}{Make plot interactive using plotly}

\item{x_label}{Label x-axis}

\item{y_label}{Label x-axis}

\item{title_label}{Title of the plot}
}
\value{
ggplot object or plotly object
}
\description{
It plots the query spectra w.r.t the reference spectra
}
\examples{
plot_spectrum_similarity <- function (spec_query, spec_ref, mz_tolerence_unit = "ppm", 
                                      mz_tolerence = 10, mz_range = c(0, 1600))
}
