% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_msp.R
\name{read_msp}
\alias{read_msp}
\title{read_msp}
\usage{
read_msp(msp_path = NULL, compound = NULL)
}
\arguments{
\item{msp_path}{The path to msp file}

\item{compound}{A vector of features/compounds to search in the msp file which is optional}
}
\value{
A list of msp data
}
\description{
It reads the msp format.
}
\examples{
read_msp(msp_path)
}
