% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_names_by_names_vector.R
\name{replace_names_by_names_vector}
\alias{replace_names_by_names_vector}
\title{replace_names_by_names_vector}
\usage{
replace_names_by_names_vector(
  data_df = NULL,
  names_vector = NULL,
  replace_in = NULL
)
}
\arguments{
\item{data_df}{The data where names to be replaced}

\item{names_vector}{The named vector used for replacing names}

\item{replace_in}{the column where names are present (it can be a column or "by_columns")}
}
\value{
The named vector of samplenames using make.names
}
\description{
It returns a dataframe after replacing names using names vector.
}
