% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_identified_from_unidentified.R
\name{separate_identified_from_unidentified}
\alias{separate_identified_from_unidentified}
\title{separate_identified_from_unidentified}
\usage{
separate_identified_from_unidentified(maven_output_df = NULL)
}
\arguments{
\item{maven_output_df}{The peak detailed format of maven output}
}
\value{
A list of two dataframes having identified and unidentified features information
}
\description{
It separates identified features from unidentified features in elmaven
}
\examples{
separate_identified_from_unidentified(maven_output_df)
}
