% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smartbind_dataframes.R
\name{smartbind_dataframes}
\alias{smartbind_dataframes}
\title{smartbind_dataframes}
\usage{
smartbind_dataframes(df_1 = NULL, df_2 = NULL)
}
\arguments{
\item{df_1}{The dataframe 1}

\item{df_2}{The dataframe 2}
}
\value{
The combined dataframe
}
\description{
It combines two dataframe having same columns
}
\examples{
smartbind_dataframes(df_1, df_2)
}
