% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_label_in_emdb.R
\name{update_label_in_emdb}
\alias{update_label_in_emdb}
\title{update_label_in_emdb}
\usage{
update_label_in_emdb(
  emdb_path = NULL,
  table_group_id = NULL,
  table_name = NULL,
  label_group_id = NULL
)
}
\arguments{
\item{emdb_path}{The path to emDB file}

\item{table_group_id}{The group id of desired metabolite}

\item{table_name}{The peak table name where desired metabolite is present}

\item{label_group_id}{A label of that feature, select from (good, bad, maybe)}
}
\value{
Update label in emDB file
}
\description{
It updates the label of feature in emDB file.
}
\examples{
update_label_in_emdb(emdb_path, table_group_id, table_name, label_group_id)
}
