% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_filtered_emdb.R
\name{write_filtered_emdb}
\alias{write_filtered_emdb}
\title{write_filtered_emdb}
\usage{
write_filtered_emdb(
  emdb_path = NULL,
  table_name = NULL,
  filtered_groups_df = NULL,
  save_path = NULL
)
}
\arguments{
\item{emdb_path}{The path to emDB file}

\item{table_name}{The peak table name where desired metabolite is present}

\item{filtered_groups_df}{A dataframe with filtered groupids}

\item{save_path}{The path to saved filtered emDB file}
}
\value{
A file is saved at save_path
}
\description{
It filters the emDB by groupIds and saves in new emDB file.
}
\examples{
get_peak_detailed_from_emdb(emdb_path, filtered_groups_df, table_name)
}
