% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_msp.R
\name{write_msp}
\alias{write_msp}
\title{write_msp}
\usage{
write_msp(msp_data = NULL, save_path = NULL)
}
\arguments{
\item{msp_data}{A list of compounds msp data which is the output of read_msp}

\item{save_path}{File path to save the data as .msp format}
}
\value{
The file saved as .msp format
}
\description{
It writes the compounds msp data into a text file with .msp extention
}
\examples{
write_msp(msp_data, save_path)
}
