% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_elements_from_msp.R
\name{get_elements_from_msp}
\alias{get_elements_from_msp}
\title{get_elements_from_msp}
\usage{
get_elements_from_msp(
  msp_path = NULL,
  compound_key = NULL,
  ignore_case = FALSE,
  load_on_memory = TRUE
)
}
\arguments{
\item{msp_path}{The path to msp file}

\item{compound_key}{A key name from msp file where features/compounds are present}

\item{ignore_case}{Ignore case while searching compound key}

\item{load_on_memory}{Load whole text into a list if TRUE else read lines one by one to save memory}
}
\value{
A vector of features/compounds
}
\description{
It converts the string text of a single compound msp into the list of msp data for a compound
}
\examples{
get_elements_from_msp(msp_path, compound_key)
}
