% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_eic.R
\name{plot_eic}
\alias{plot_eic}
\title{plot_eic}
\usage{
plot_eic(
  intensity_data = NULL,
  rt_min = NULL,
  rt_max = NULL,
  x_label = "RT",
  y_label = "Intensity",
  title_label = "",
  legend_label = "Sample",
  x_label_size = 16,
  y_label_size = 16,
  title_label_size = 16,
  legend_label_size = 16,
  x_text_size = 14,
  y_text_size = 14,
  legend_text_size = 14,
  interactive = TRUE
)
}
\arguments{
\item{intensity_data}{The intensity data with sample, intensity and rt columns}

\item{rt_min}{Make dashed line at rt_min value}

\item{rt_max}{Make dashed line at rt_max value}

\item{x_label}{Label x-axis}

\item{y_label}{Label x-axis}

\item{title_label}{Title of the plot}

\item{legend_label}{Label for the lagends}

\item{x_label_size}{Text size of x label}

\item{y_label_size}{Text size of y label}

\item{title_label_size}{Text size of text label}

\item{legend_label_size}{Text size of legend label}

\item{x_text_size}{Text size of x ticks}

\item{y_text_size}{Text size of y ticks}

\item{legend_text_size}{Text size of legends}

\item{interactive}{Make plot interactive using plotly}
}
\value{
ggplot object or plotly object
}
\description{
It plots the extracted ion chromatogram (EIC) for provided intensity data.
}
\examples{
plot_eic(intensity_data, rt_min, rt_min)
}
