% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spectrum_similarity.R
\name{plot_spectrum_similarity}
\alias{plot_spectrum_similarity}
\title{plot_spectrum_similarity}
\usage{
plot_spectrum_similarity(
  spec_query = NULL,
  spec_ref = NULL,
  mz_tolerence = 20,
  mz_tolerence_unit = "ppm",
  annotate_mz = FALSE,
  annotate_mz_cutoff = NULL,
  intensity_cutoff = NULL,
  mz_precision = NULL,
  mz_range = NULL,
  x_label = NULL,
  y_label = NULL,
  title_label = NULL,
  top_label = NULL,
  bottom_label = NULL,
  interactive = FALSE
)
}
\arguments{
\item{spec_query}{The query spectrum dataframe with mz and intensity columns}

\item{spec_ref}{The reference spectrum dataframe with mz and intensity columns}

\item{mz_tolerence}{Value of mz tolerence}

\item{mz_tolerence_unit}{The mz tolerence unit (ppm or Da)}

\item{annotate_mz}{Annotate mz values on spectrum, select from TRUE/FALSE}

\item{annotate_mz_cutoff}{The intensity cutoff in percentage w.r.t base peak (highest intensity peak) used to annotate mz peaks}

\item{intensity_cutoff}{The intensity cutoff in percentage w.r.t base peak (highest intensity peak) used to filter data}

\item{mz_precision}{The numeric precision of mz which is equal to the number of digits to show on mz annotation}

\item{mz_range}{The mz range used to calculate score}

\item{x_label}{Label x-axis}

\item{y_label}{Label x-axis}

\item{title_label}{Title of the plot}

\item{top_label}{Specify top label}

\item{bottom_label}{Specify bottom label}

\item{interactive}{Make plot interactive using plotly}
}
\value{
ggplot object or plotly object
}
\description{
It plots the query spectra w.r.t the reference spectra
}
\examples{
plot_spectrum_similarity(spec_query, spec_ref, mz_tolerence = 10, mz_tolerence_unit = "ppm")
}
