% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_msp.R
\name{read_msp}
\alias{read_msp}
\title{read_msp}
\usage{
read_msp(
  msp_path = NULL,
  compound = NULL,
  compound_key = "Name",
  exact_match = TRUE
)
}
\arguments{
\item{msp_path}{The path to msp file}

\item{compound}{A vector of features/compounds to search in the msp file which is optional}

\item{compound_key}{A key name from msp file where features/compounds are present}

\item{exact_match}{Search exact match (TRUE) or do partial matching (FALSE) of compound names in the msp file}
}
\value{
A list of msp data
}
\description{
It reads the msp format.
}
\examples{
read_msp(msp_path)
}
