% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restructure_camera_annotations.R
\name{restructure_camera_annotations}
\alias{restructure_camera_annotations}
\title{restructure_camera_annotations}
\usage{
restructure_camera_annotations(
  annotated_peaks_df = NULL,
  polarity = "positive"
)
}
\arguments{
\item{annotated_peaks_df}{The camera output}

\item{polarity}{The polarity of the data}
}
\value{
A dataframe with restructure camera output
}
\description{
It restructures the camera output to group features
}
\examples{
restructure_camera_annotations(annotated_peaks_df, polarity = 'positive')
}
